add a new annotation to append custom ignored paths
the deployment annotation `osiris.deislabs.io/ignoredPaths` can be used to append custom paths
to the list of "ignored paths" - that won't be counted in the metrics.

this allow to configure a platform-wide list of ignoredPaths in osiris configuration (helm chart for example),
and to add more paths on a per-deployment basis.

same as for the global configuration, the list of paths should be written as a comma-separated string.